requirements.txt
crewai==0.95.0
streamlit==1.48.0
pandas==2.2.3
numpy==2.2.4
langchain-google-genai==2.1.8
google-generativeai==0.8.5
langchain==0.3.25
langchain-core==0.3.70
matplotlib==3.10.1
seaborn==0.13.2
plotly==5.17.0
reportlab==4.0.7
python-dotenv==1.0.1
typing-extensions==4.12.2
scipy==1.15.2
scikit-learn==1.6.1
openpyxl==3.1.5
